Implement a rule to check deprecated hook implementations
This will analyse hook implementations in `.inc` and `.module` files. It
detects a hook by looking at the pattern `mymodule_[hook]` which is
detected from the file-name the function is in.

It will replace `mymodule` with `hook` and then try to find a function
with that name. For this to work PHPStan Drupal's `DrupalAutoloader`
class must be adjusted to load `.api.php` files as well so that the
hooks can be discovered. In testing on the Open Social code-base
this did not cause any issues.

The rule will provide the deprecation message in the error if it's
provided, if no deprecation message was provided (i.e. a lonesome
`@deprecated` tag) then the hook will simply be shown as deprecated
without advice.